Refactor IteratorStream
Refactor the IteratorStream implementation of the PSR-7 StreamInterface to flip the buffering logic. Instead of building a buffer up to the size of the read() length parameter before returning the value, we return whatever we get from the iterator's current(), and only buffer the value if it exceeds the lenght. We also move where we call `next()` to avoid blocking behavior from delaying returning the current() result.